Remove (now) unused stuff
The previous patch broke the python stuff. This patch removes it completely.
Having a separate 2048.h header is now also no longer useful.
