--- 2048.h
@@ -1,29 +1,0 @@
-#include <stdlib.h>
-#include "platdefs.h"
-
-/* The fundamental trick: the 4x4 board is represented as a 64-bit word,
- * with each board square packed into a single 4-bit nibble.
- * 
- * The maximum possible board value that can be supported is 32768 (2^15), but
- * this is a minor limitation as achieving 65536 is highly unlikely under normal circumstances.
- * 
- * The space and computation savings from using this representation should be significant.
- * 
- * The nibble shift can be computed as (r,c) -> shift (4*r + c). That is, (0,0) is the LSB.
- */
-
-typedef uint64_t board_t;
-typedef uint16_t row_t;
-
-static const board_t ROW_MASK = 0xFFFFULL;
-
-static void print_board(board_t board) {
-    for (int i = 0; i < 4; ++i) {
-        for (int j = 0; j < 4; ++j) {
-            putchar("0123456789abcdef"[board & 0xf]);
-            board >>= 4;
-        }
-        putchar('\n');
-    }
-    putchar('\n');
-}
